move math/fft -> utils/fft, fri, simplify channels, move oods
