+++ src/channel_commit/bitcoin_script.rs
@@ -14,20 +14,6 @@ impl Pushable for Commitment {
 /// Gadget for committing field elements.
 pub struct CommitmentGadget;
 impl CommitmentGadget {
-    /// Commit a m31 element.
-    pub fn commit_m31() -> Script {
-        script! {
-            OP_SHA256
-        }
-    }
-
-    /// Commit a cm31 element.
-    pub fn commit_cm31() -> Script {
-        script! {
-            OP_SHA256 OP_CAT OP_SHA256
-        }
-    }
-
     /// Commit a qm31 element.
     pub fn commit_qm31() -> Script {
         script! {
@@ -46,50 +32,6 @@ mod test {
     use stwo_prover::core::fields::cm31::CM31;
     use stwo_prover::core::fields::m31::M31;
     use stwo_prover::core::fields::qm31::QM31;
-
-    #[test]
-    fn test_commit_m31() {
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-
-        let commit_script = CommitmentGadget::commit_m31();
-        println!("M31.commit() = {} bytes", commit_script.len());
-
-        for _ in 0..100 {
-            let a = M31::reduce(prng.next_u64());
-            let b = Commitment::commit_m31(a);
-
-            let script = script! {
-                { a }
-                { commit_script.clone() }
-                { b.clone() }
-                OP_EQUAL
-            };
-            let exec_result = execute_script(script);
-            assert!(exec_result.success);
-        }
-    }
-
-    #[test]
-    fn test_commit_cm31() {
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-
-        let commit_script = CommitmentGadget::commit_cm31();
-        println!("CM31.commit() = {} bytes", commit_script.len());
-
-        for _ in 0..100 {
-            let a = CM31(M31::reduce(prng.next_u64()), M31::reduce(prng.next_u64()));
-            let b = Commitment::commit_cm31(a);
-
-            let script = script! {
-                { a }
-                { commit_script.clone() }
-                { b.clone() }
-                OP_EQUAL
-            };
-            let exec_result = execute_script(script);
-            assert!(exec_result.success);
-        }
-    }
 
     #[test]
     fn test_commit_qm31() {
