+++ src/circle/bitcoin_script.rs
@@ -1,15 +1,8 @@
 use crate::treepp::*;
-use num_traits::One;
 use rust_bitcoin_m31::{
-    m31_add_n31, m31_sub, push_m31_one, push_n31_one, push_qm31_one, qm31_add, qm31_copy,
-    qm31_double, qm31_dup, qm31_equalverify, qm31_from_bottom, qm31_fromaltstack, qm31_mul,
-    qm31_neg, qm31_roll, qm31_rot, qm31_square, qm31_sub, qm31_swap, qm31_toaltstack,
+    push_qm31_one, qm31_add, qm31_copy, qm31_double, qm31_equalverify, qm31_fromaltstack, qm31_mul,
+    qm31_roll, qm31_square, qm31_sub, qm31_swap, qm31_toaltstack,
 };
-use std::ops::{Add, Mul, Neg};
-use stwo_prover::core::fields::qm31::QM31;
-use stwo_prover::core::fields::{Field, FieldExpOps};
-
-use crate::channel::ChannelGadget;
 
 /// Gadget for points on the circle curve in the qm31 field.
 pub struct CirclePointGadget;
@@ -82,92 +75,23 @@ impl CirclePointGadget {
             qm31_sub
         }
     }
-
-    /// Samples a random point over the projective line, see Lemma 1 in https://eprint.iacr.org/2024/278.pdf
-    ///
-    /// hint:
-    ///  w - qm31 hint (5 elements)
-    ///
-    /// input:
-    ///  x - (1-t^2)/(1+t^2), where t is extracted from channel (4 elements)
-    ///  y - 2t/(1+t^2), where t is extracted from channel (4 elements)
-    ///  channel
-    ///
-    /// output:
-    ///  channel'=sha256(channel)
-    ///  x
-    ///  y
-    /// where (x,y) - random point on C(QM31) satisfying x^2+y^2=1 (8 elements)
-    pub fn get_random_point() -> Script {
-        script! {
-            { ChannelGadget::squeeze_qm31_using_hint() }
-            // stack: x, y, channel', t
-
-            // compute t^2 from t
-            qm31_dup
-            qm31_square
-
-            // compute t^2 - 1
-            qm31_dup
-            push_m31_one
-            m31_sub // a trick
-
-            // compute t^2 + 1
-            qm31_swap
-            push_n31_one
-            m31_add_n31
-            qm31_dup
-
-            // stack: x, y, channel', t, t^2 - 1, t^2 + 1, t^2 + 1
-
-            // pull the hint x and verify
-            qm31_from_bottom
-            qm31_dup
-            qm31_rot
-            qm31_mul
-            qm31_neg
-            { qm31_roll(3) }
-            qm31_equalverify
-
-            // stack: y, channel', t, t^2 + 1, x
-
-            // pull the hint y
-            qm31_from_bottom
-            qm31_dup
-            { qm31_roll(3) }
-            qm31_mul
-            { qm31_roll(3) }
-            qm31_double
-            qm31_equalverify
-        }
-    }
-
-    /// Push the hint for sampling a random circle curve point over qm31.
-    pub fn push_random_point_hint(t: QM31) -> Script {
-        let one_plus_tsquared_inv = t.square().add(QM31::one()).inverse();
-
-        script! {
-            { QM31::one().add(t.square().neg()).mul(one_plus_tsquared_inv) } // x = (1 - t^2) / (1 + t^2)
-            { t.double().mul(one_plus_tsquared_inv) } // y = 2t / (1 + t^2)
-        }
-    }
 }
 
 #[cfg(test)]
 mod test {
     use num_traits::One;
-    use std::ops::{Add, Mul, Neg};
+    use std::ops::{Add, Neg};
     use stwo_prover::core::circle::CirclePoint;
 
     use crate::treepp::*;
-    use rand::{Rng, RngCore, SeedableRng};
+    use rand::{RngCore, SeedableRng};
     use rand_chacha::ChaCha20Rng;
     use rust_bitcoin_m31::qm31_equalverify;
     use stwo_prover::core::fields::m31::M31;
     use stwo_prover::core::fields::qm31::QM31;
     use stwo_prover::core::fields::{Field, FieldExpOps};
 
-    use crate::{channel::Channel, channel_extract::ExtractorGadget, circle::CirclePointGadget};
+    use crate::circle::CirclePointGadget;
 
     #[test]
     fn test_add() {
@@ -272,51 +196,5 @@ mod test {
             let exec_result = execute_script(script);
             assert!(exec_result.success);
         }
-    }
-
-    #[test]
-    fn test_get_random_point() {
-        let mut prng = ChaCha20Rng::seed_from_u64(0);
-
-        let get_random_point_script = CirclePointGadget::get_random_point();
-        println!(
-            "CirclePointSecure.get_random_point() = {} bytes",
-            get_random_point_script.len()
-        );
-
-        let mut a = [0u8; 32];
-        a.iter_mut().for_each(|v| *v = prng.gen());
-
-        let mut channel = Channel::new(a);
-        let (t, hint_t) = channel.draw_qm31();
-
-        let c = channel.state;
-
-        let x = t
-            .square()
-            .add(QM31::one())
-            .inverse()
-            .mul(QM31::one().add(t.square().neg())); //(1+t^2)^-1 * (1-t^2)
-        let y = t
-            .square()
-            .add(QM31::one())
-            .inverse()
-            .mul(QM31::one().double().mul(t)); // (1+t^2)^-1 * 2 * t
-
-        let script = script! {
-            { ExtractorGadget::push_hint_qm31(&hint_t) }
-            { CirclePointGadget::push_random_point_hint(t) }
-            { a.to_vec() }
-            { get_random_point_script.clone() }
-            { y } // check y
-            qm31_equalverify
-            { x } // check x
-            qm31_equalverify
-            { c.to_vec() } // check channel'
-            OP_EQUALVERIFY // checking that indeed channel' = sha256(channel)
-            OP_TRUE
-        };
-        let exec_result = execute_script(script);
-        assert!(exec_result.success);
     }
 }
