--- src/math/mod.rs
@@ -1,5 +1,0 @@
-mod bitcoin_script;
-pub use bitcoin_script::*;
-
-/// Module for the FFT implementation.
-pub mod fft;
