+++ src/twiddle_merkle_tree/mod.rs
@@ -1,4 +1,4 @@
-use crate::math::fft::get_twiddles;
+use crate::utils::get_twiddles;
 use crate::utils::num_to_bytes;
 use sha2::{Digest, Sha256};
 use stwo_prover::core::fields::m31::M31;
